# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boost & Earn API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  fullName: string;
  phone: string;
  password: string;
  referralCode?: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface User {
  id: number;
  fullName: string;
  phone: string;
  balance: number;
  referralCode: string;
  isAdmin: boolean;
  isSuspended: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface AdminUser {
  id: number;
  fullName: string;
  phone: string;
  balance: number;
  referralCode: string;
  isAdmin: boolean;
  isSuspended: boolean;
  teamCount: number;
  createdAt: string;
}

export interface Task {
  id: number;
  link: string;
  platform: string;
  taskName: string;
  durationDays: number;
  expiresAt: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateTaskInput {
  link: string;
  platform: string;
  taskName: string;
  durationDays: number;
}

export interface UpdateTaskInput {
  link?: string;
  platform?: string;
  taskName?: string;
  durationDays?: number;
  isActive?: boolean;
}

export interface PlatformCounts {
  tiktok: number;
  facebook: number;
  youtube: number;
  instagram: number;
}

export interface Submission {
  id: number;
  userId: number;
  taskId: number;
  imageUrls: string[];
  status: string;
  rejectionReason?: string;
  userName?: string;
  taskName?: string;
  platform?: string;
  createdAt: string;
}

export interface CreateSubmissionInput {
  taskId: number;
}

export interface RejectInput {
  reason: string;
}

export interface Withdrawal {
  id: number;
  userId: number;
  amount: number;
  phone: string;
  name: string;
  operator: string;
  status: string;
  userName?: string;
  createdAt: string;
}

export interface CreateWithdrawalInput {
  amount: number;
  phone: string;
  name: string;
  operator: string;
}

export interface Notification {
  id: number;
  userId: number;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

export type AdjustBalanceInputOperation =
  (typeof AdjustBalanceInputOperation)[keyof typeof AdjustBalanceInputOperation];

export const AdjustBalanceInputOperation = {
  add: "add",
  subtract: "subtract",
} as const;

export interface AdjustBalanceInput {
  amount: number;
  operation: AdjustBalanceInputOperation;
}

export interface ResetPasswordInput {
  newPassword: string;
}

export interface Deposit {
  id: number;
  userId: number;
  link: string;
  platform: string;
  taskType: string;
  durationDays: number;
  paymentProofUrl: string;
  status: string;
  userName?: string;
  createdAt: string;
}

export interface CreateDepositInput {
  link: string;
  platform: string;
  taskType: string;
  durationDays: number;
}

export interface ApproveDepositInput {
  taskName: string;
  durationDays: number;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
}

export interface TeamMember {
  id: number;
  fullName: string;
  phone: string;
  balance: number;
  createdAt: string;
}

export type GetTasksParams = {
  platform?: string;
};