# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boost & Earn API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/balance:
    post:
      operationId: adjustUserBalance
      tags: [admin]
      summary: Adjust user balance
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceInput"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/suspend:
    post:
      operationId: toggleUserSuspension
      tags: [admin]
      summary: Suspend or unsuspend user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspension toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [admin]
      summary: Reset user password
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /deposits:
    get:
      operationId: getDeposits
      tags: [admin]
      summary: Get all deposits (announcer submissions)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [admin]
      summary: Submit a deposit (announcer)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateDepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{depositId}/approve:
    post:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve deposit and publish task
      security:
        - bearerAuth: []
      parameters:
        - name: depositId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDepositInput"
      responses:
        "200":
          description: Deposit approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /referrals/link:
    get:
      operationId: getReferralLink
      tags: [users]
      summary: Get user referral link
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
  /referrals/team:
    get:
      operationId: getTeam
      tags: [users]
      summary: Get user's referral team
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        password:
          type: string
        referralCode:
          type: string
      required:
        - fullName
        - phone
        - password
    LoginInput:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        balance:
          type: number
        referralCode:
          type: string
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - phone
        - balance
        - referralCode
        - isAdmin
        - isSuspended
        - createdAt
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        balance:
          type: number
        referralCode:
          type: string
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        teamCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - fullName
        - phone
        - balance
        - referralCode
        - isAdmin
        - isSuspended
        - teamCount
        - createdAt
...
    AdjustBalanceInput:
      type: object
      properties:
        amount:
          type: number
        operation:
          type: string
          enum: [add, subtract]
      required:
        - amount
        - operation
    ResetPasswordInput:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        link:
          type: string
        platform:
          type: string
        taskType:
          type: string
        durationDays:
          type: integer
        paymentProofUrl:
          type: string
        status:
          type: string
        userName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - link
        - platform
        - taskType
        - durationDays
        - paymentProofUrl
        - status
        - createdAt
    CreateDepositInput:
      type: object
      properties:
        link:
          type: string
        platform:
          type: string
        taskType:
          type: string
        durationDays:
          type: integer
      required:
        - link
        - platform
        - taskType
        - durationDays
    ApproveDepositInput:
      type: object
      properties:
        taskName:
          type: string
        durationDays:
          type: integer
      required:
        - taskName
        - durationDays
    ReferralInfo:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        totalReferrals:
          type: integer
      required:
        - referralCode
        - referralLink
        - totalReferrals
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        balance:
          type: number
        createdAt:
          type: string
      required:
        - id
        - fullName
        - phone
        - balance
        - createdAt