# lib/api-zod/src/generated/types/adjustBalanceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boost & Earn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdjustBalanceInputOperation } from "./adjustBalanceInputOperation";

export interface AdjustBalanceInput {
  amount: number;
  operation: AdjustBalanceInputOperation;
}