# lib/api-zod/src/generated/types/referralInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boost & Earn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
}