# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boost & Earn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  phone: string;
  balance: number;
  referralCode: string;
  isAdmin: boolean;
  isSuspended: boolean;
  createdAt: string;
}